requirements.txt
fastapi[standard]
pymongo
python-dotenv
python-multipart
bcrypt
pyjwt
uvicorn
pydantic
